# Conflicts:
#	src/main/java/com/example/peeppo/domain/auction/service/AuctionService.java
#	src/main/java/com/example/peeppo/domain/user/service/UserService.java

+++ src/main/java/com/example/peeppo/domain/goods/service/GoodsService.java
@@ -42,10 +42,7 @@ import org.springframework.web.multipart.MultipartFile;
 import org.springframework.web.util.UriUtils;
 
 import java.time.LocalDateTime;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Objects;
-import java.util.Optional;
+import java.util.*;
 import java.util.stream.Collectors;
 
 import static com.example.peeppo.domain.goods.enums.GoodsStatus.SOLDOUT;
@@ -85,6 +82,7 @@ public class GoodsService {
             throw new IllegalArgumentException("물품 이미지가 꼭 필요합니다.");
         }
         WantedGoods wantedGoods = new WantedGoods(wantedRequestDto);
+        System.out.println("실행체크");
         Goods goods = new Goods(goodsRequestDto, wantedGoods, user, GoodsStatus.ONSALE);
         RatingGoods ratingGoods = new RatingGoods(goods);
 
@@ -173,9 +171,9 @@ public class GoodsService {
             goodsRecent.remove(0);
         }
         // goodsRecent.add(Long.toString(goods.getGoodsId())); // 조회시에 리스트에 추가 !
-        Optional<Dibs> dibsGoods = dibsRepository.findByUserUserIdAndGoodsGoodsId(user.getUserId(), goodsId);
+        Optional<Dibs> dibsGoods = dibsRepository.findByUserUserIdAndGoodsGoodsIdAndGoodsIsDeletedFalse(user.getUserId(), goodsId);
         boolean checkDibs = dibsGoods.isPresent();
-        Long dibsCount = dibsRepository.countByGoodsGoodsId(goodsId);
+        Long dibsCount = dibsRepository.countByGoodsGoodsIdAndGoodsIsDeletedFalse(goodsId);
         return new ApiResponse<>(true, new GoodsDetailResponseDto(new GoodsResponseDto(goods, imageUrls, wantedGoods, checkSameUser, checkDibs, dibsCount), rcGoodsResponseDtoList), null);
     }
 
@@ -233,13 +231,19 @@ public class GoodsService {
 
     @Transactional
     public ApiResponse<DeleteResponseDto> deleteGoods(Long goodsId, Long userId) throws IllegalAccessException {
-        userRatingHelper.getUser(userId);
         Goods goods = findGoods(goodsId);
         if (Objects.equals(userId, goods.getUser().getUserId())) {
-            goods.delete();
-            goodsRepository.save(goods);
+            if (goods.getGoodsStatus().equals(GoodsStatus.ONSALE) ||
+                    goods.getGoodsStatus().equals(GoodsStatus.SOLDOUT)) {
+
+                goods.changeStatus(GoodsStatus.ONSALE);
+                goods.delete();
+                goodsRepository.save(goods);
+            } else {
+                throw new IllegalArgumentException("입찰, 경매, 거래중인 상품은 삭제할 수 없습니다.");
+            }
         } else {
-            throw new IllegalAccessException();
+            throw new IllegalAccessException("유저의 정보가 올바르지 않습니다.");
         }
         return new ApiResponse<>(true, new DeleteResponseDto("삭제되었습니다"), null);
     }
@@ -315,7 +319,7 @@ public class GoodsService {
     }
 
     public ApiResponse<List<GoodsListResponseDto>> searchGoods(String keyword) {
-        List<Goods> searchList = goodsRepository.findByTitleContaining(keyword);
+        List<Goods> searchList = goodsRepository.findByTitleContainingAndIsDeletedFalse(keyword);
         List<GoodsListResponseDto> goodsListResponseDtos = new ArrayList<>();
         for (Goods search : searchList) {
             goodsListResponseDtos.add(new GoodsListResponseDto(search));
@@ -367,7 +371,6 @@ public class GoodsService {
     }
 
 
-
     // 내가 보낸 교환요청
     public ResponseEntity<Page<GoodsRequestResponseDto>> requestTradeList(User user, int page, int size, String sortBy, boolean isAsc,
                                                                           String requestStatusStr) {
@@ -413,27 +416,27 @@ public class GoodsService {
         RequestStatus requestStatus1;
         userRatingHelper.getUser(user.getUserId());
 
-       List<GoodsRequestResponseDto> goodsRequestResponseDtos = new ArrayList<>();
+        List<GoodsRequestResponseDto> goodsRequestResponseDtos = new ArrayList<>();
 
-        if(requestStatusStr != null) {
-            //1) requestgoods 테이블에서 seller goods 전부 찾아오기 => DTO 변환해주기
+        if (requestStatusStr != null) {
+//1) requestgoods 테이블에서 seller goods 전부 찾아오기 => DTO 변환해주기
             requestStatus1 = RequestStatus.valueOf(requestStatusStr);
             requestGoods = requestRepository.findByReceiveUserAndRequestStatus(user.getUserId(), requestStatus1, pageable);
         } else {
             requestGoods = requestRepository.findByReceiveUser(user.getUserId(), pageable);
         }
         List<GoodsRequestResponseDto> finalResponseDto = new ArrayList<>();
-        //2) requestGoods 순회하면서 buyerGoods 찾아오기
+//2) requestGoods 순회하면서 buyerGoods 찾아오기
 
         for (User buyer : requestGoods) {
-           List<Goods> requestGoods1 = requestRepository.findByBuyerUserAndSeller(buyer.getUserId(), user.getUserId()); //seller의 굿즈를 찾아왔어
-            for(Goods goods : requestGoods1){
+            List<Goods> requestGoods1 = requestRepository.findByBuyerUserAndSeller(buyer.getUserId(), user.getUserId()); //seller의 굿즈를 찾아왔어
+            for (Goods goods : requestGoods1) {
                 LocalDateTime createAt = null;
                 RequestStatus requestStatus = null;
                 List<RequestGoods> requestGoods2 = requestRepository.findAllBySellerGoodsIdAndUserId(goods.getGoodsId(), buyer.getUserId());
                 RequestSingleResponseDto goodsListResponseDto = new RequestSingleResponseDto(goods);
                 List<RequestSingleResponseDto> goodsListResponseDtos = new ArrayList<>(); // 같은 물건에 요청 넣은 친구들 저장해서 넣어줌 !
-                for(RequestGoods requestGoods3 : requestGoods2){
+                for (RequestGoods requestGoods3 : requestGoods2) {
                     createAt = requestGoods3.getCreatedAt();
                     requestStatus = requestGoods3.getRequestStatus();
                     RequestSingleResponseDto goodsListResponseDtos2 = new RequestSingleResponseDto(requestGoods3.getBuyer());
@@ -448,20 +451,18 @@ public class GoodsService {
 
     //내물건이 아니여야한다 !
     @Transactional
-    public ResponseDto goodsRequest (User user, GoodsRequestRequestDto goodsRequestRequestDto, Long urGoodsId){
+    public ResponseDto goodsRequest(User user, GoodsRequestRequestDto goodsRequestRequestDto, Long urGoodsId) {
         Goods urGoods = goodsRepository.findByGoodsId(urGoodsId) // sellergoods(남의 물건)
                 .orElseThrow(() -> new NullPointerException("해당 상품이 존재하지 않습니다."));
-        List<RequestGoods> requestGoods = new ArrayList<>();
+                List<RequestGoods> requestGoods = new ArrayList<>();
         List<Goods> goodsList = new ArrayList<>();
-
-        RequestGoods requestGoods1 = requestRepository.findBySellerAndUserUserId(urGoods, user.getUserId());
-        if(requestGoods1 != null){
-            throw new IllegalArgumentException("해당 물품과 교환 요청 중입니다. 취소후 다시 시도주세요.");
-        }
 
         for (Long goodsId : goodsRequestRequestDto.getGoodsId()) {
             Goods goods = goodsRepository.findById(goodsId).orElseThrow(
                     () -> new IllegalArgumentException("존재하지 않는 goodsId 입니다.")); // 내 물건
+            if(!goods.getUser().getUserId().equals(user.getUserId())){
+                throw new IllegalArgumentException("자신의 물품만 교환요청 할 수 있습니다.");
+            }
 
 
             if (!(urGoods.getUser().equals(goods.getUser()))) {
@@ -489,7 +490,7 @@ public class GoodsService {
         for (Goods goods : goodsPage.getContent()) {
             boolean checkSameUser = Objects.equals(goods.getUser().getUserId(), user.getUserId());
             Image image = imageRepository.findByGoodsGoodsIdOrderByCreatedAtAscFirst(goods.getGoodsId());
-            boolean checkDibs = dibsRepository.findByUserUserIdAndGoodsGoodsId(user.getUserId(), goods.getGoodsId()).isPresent();
+            boolean checkDibs = dibsRepository.findByUserUserIdAndGoodsGoodsIdAndGoodsIsDeletedFalse(user.getUserId(), goods.getGoodsId()).isPresent();
             goodsResponseList.add(new GoodsListResponseDto(goods, image.getImageUrl(), checkDibs, checkSameUser));
         }
         return new PageResponse<>(goodsResponseList, pageable, goodsPage.getTotalElements());
@@ -516,7 +517,8 @@ public class GoodsService {
 
     public ResponseDto ratingCheck(User user) {
         userRatingHelper.getUser(user.getUserId());
-        Goods goods = goodsRepository.findByUserUserId(user.getUserId());
+        Goods goods = goodsRepository.findByUserUserId(user.getUserId())
+                .orElseThrow(() -> new NullPointerException("존재하지 않는 상품입니다."));
         goods.changeCheck(true);
         goodsRepository.save(goods);
 
@@ -555,17 +557,28 @@ public class GoodsService {
     }
 
     // 승인 => 교환중으로 상태 변경
-    public void goodsAccept(RequestAcceptRequestDto requestAcceptRequestDto, User user) {
-        for (Long goodsId : requestAcceptRequestDto.getRequestId()) {
-            RequestGoods requestGoods = requestRepository.findByBuyerGoodsId(goodsId)
-                    .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 게시물입니다"));
+    public void goodsAccept(Long sellerGoodsId, RequestAcceptRequestDto requestAcceptRequestDto, User user) {
+        List<RequestGoods> buyerRequest = requestRepository.findAllBySellerGoodsId(sellerGoodsId);
+        List<Goods> goodsList = new ArrayList<>();
+
+        for (RequestGoods requestGoods : buyerRequest) {
             if (!requestGoods.getSeller().getUser().getUserId().equals(user.getUserId())) {
                 throw new IllegalArgumentException("물품 교환 요청 수락은 본인만 가능합니다.");
             }
-            requestGoods.changeStatus(RequestStatus.TRADING);
-            requestRepository.save(requestGoods);
+            Goods goods = requestGoods.getBuyer();
+            if (requestAcceptRequestDto.getRequestId().contains(requestGoods.getBuyer().getGoodsId())) {
+                requestGoods.changeStatus(RequestStatus.TRADING);
+            } else {
+                requestGoods.changeStatus(RequestStatus.CANCEL);
+                goods.changeStatus(GoodsStatus.ONSALE);
+                goodsList.add(goods);
+            }
         }
+
+        requestRepository.saveAll(buyerRequest);
+        goodsRepository.saveAll(goodsList);
     }
+
 
     // 요청 -> 거절
     public void goodsRefuse(RequestAcceptRequestDto requestAcceptRequestDto, User user) {
@@ -587,11 +600,11 @@ public class GoodsService {
         List<RequestGoods> requestGoodsList = new ArrayList<>();
         for (Long requestGoodsId : requestAcceptRequestDto.getRequestId()) {
 
-            RequestGoods requestGoods = requestRepository.findBuyerGoodsId(requestGoodsId);
+            RequestGoods requestGoods = requestRepository.findBuyerGoodsIdAndUserId(requestGoodsId, userDetails.getUser().getUserId(), RequestStatus.TRADING);
             if (!(requestGoods.getRequestStatus().equals(RequestStatus.TRADING))) {
                 throw new IllegalArgumentException("정상적인 접근이 아닙니다.");
             }
-            if (!(Objects.equals(requestGoods.getSeller().getUser(), userDetails.getUser()))) {
+            if (!(Objects.equals(requestGoods.getSeller().getUser().getUserId(), userDetails.getUser().getUserId()))) {
                 throw new IllegalArgumentException("판매 물품의 주인이 아니라면 교환 완료를 진행할 수 없습니다...");
             }
             buyerAndSellerList.add(requestGoods.getBuyer());
@@ -608,27 +621,27 @@ public class GoodsService {
         return new ApiResponse<>(true, new MsgResponseDto("상대방의 교환완료를 기다리는중..."), null);
     }
 
-  /*  @Transactional
-    public ApiResponse<MsgResponseDto> tradeCompleted(RequestAcceptRequestDto requestAcceptRequestDto,
-                                                      UserDetailsImpl userDetails) {
-        List<RequestGoods> requestGoodsList = new ArrayList<>();
-        List<Long> buyerGoodsIds = new ArrayList<>();
-        Long userId = userDetails.getUser().getUserId();
-        for (int i = 0; i < requestAcceptRequestDto.getRequestId().size(); i++) {
-            RequestGoods requestGoods = requestRepository.findByBuyerGoodsId(requestAcceptRequestDto.getRequestId().get(i))
-                    .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 게시물입니다"));
-            if (!requestGoods.getRequestStatus().equals(RequestStatus.TRADING)) {
-                throw new IllegalArgumentException("정상적인 접근이 아닙니다.");
-            }
-            if(!requestGoods.getUser().getUserId().equals(userId) ||
-            !requestGoods.getBuyer().getUser().getUserId().equals(userId)){
-                throw new IllegalArgumentException("물품 주인이 아니라면 거래를 진행할 수 있습니다..");
-            }
-            requestGoodsList.add(requestGoods);
-            buyerGoodsIds.add(requestGoods.getBuyer().getGoodsId());
-        }
-        Goods buyerGoods = requestGoodsList.get(0).getBuyer();
-        Goods sellerGoods = requestGoodsList.get(0).getSeller();
+    /*  @Transactional
+      public ApiResponse<MsgResponseDto> tradeCompleted(RequestAcceptRequestDto requestAcceptRequestDto,
+                                                        UserDetailsImpl userDetails) {
+          List<RequestGoods> requestGoodsList = new ArrayList<>();
+          List<Long> buyerGoodsIds = new ArrayList<>();
+          Long userId = userDetails.getUser().getUserId();
+          for (int i = 0; i < requestAcceptRequestDto.getRequestId().size(); i++) {
+              RequestGoods requestGoods = requestRepository.findByBuyerGoodsId(requestAcceptRequestDto.getRequestId().get(i))
+                      .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 게시물입니다"));
+              if (!requestGoods.getRequestStatus().equals(RequestStatus.TRADING)) {
+                  throw new IllegalArgumentException("정상적인 접근이 아닙니다.");
+              }
+              if(!requestGoods.getUser().getUserId().equals(userId) ||
+              !requestGoods.getBuyer().getUser().getUserId().equals(userId)){
+                  throw new IllegalArgumentException("물품 주인이 아니라면 거래를 진행할 수 있습니다..");
+              }
+              requestGoodsList.add(requestGoods);
+              buyerGoodsIds.add(requestGoods.getBuyer().getGoodsId());
+          }
+          Goods buyerGoods = requestGoodsList.get(0).getBuyer();
+          Goods sellerGoods = requestGoodsList.get(0).getSeller();
 
           if (buyerGoods.getUser().getUserId().equals(userId)) {
               List<Goods> buyerGoodsList = buyerGoodsListStatusChange(buyerGoodsIds, TRADING);
